chore(primitives): remove deref of tempo header to eth header (#436)
Removes the `Deref<Target = Header> for TempoHeader` impl. I found this
to be a footgun, where depending whether or not
`alloy_primitives::Sealable` was imported, I would get the result of
`Sealable::hash_slow(&tempo_header)` or the inherent `Header::hash_slow`
due to deref.

This not only happens when one hold a `TempoHeader`, but even if one
holds a `Block`: because of `Deref<Target = H> for Block<H>`, auto-deref
will go `Block -> TempoHeader -> Header`.