chore: make fuzz CI optional (#1263)
The fuzz CI builds tempo-foundry with latest PR branch, to check if the
tests work.
But if there is a breaking change, tempo-foundry might take some time
update this, or before merging in a fix, they would need to bump the
commit to latest main.

Which means there is a race condition. We solve this by making the fuzz
CI optional, so that we can give tooling a commit hash on main to work
on.

In all non breaking cases, a failing CI should not be merged.